image container by width
It is for text readable image.

Originally, lightbox2 resizing image function depends on screen height.
But if it has text it makes hard to read. This patch, it makes to
resize a image depends on screen width.

There is 3 option. You should turn off default fitImagesInViewport and
turn on fitImagesInViewportByWidth. You need to specify the max width.

this.fitImagesInViewport          = false;
this.fitImagesInViewportByWidth   = true;
this.containerMaxWidth            = 1140;